Moves inner product tests
